requirements.txt
pygame==2.6.1
PyOpenGL==3.1.10
PyOpenGL-accelerate==3.1.10
PyWavefront==1.3.3
noise==1.2.2
numpy>=1.21.0
Pillow==10.1.0
